Refactor command and alert handling for improved MVVM
- Updated `RelayCommand` to use `Predicate<object>?` for `canExecute` and added `CanExecuteChanged` event.
- Refactored `Alert` class to implement `INotifyPropertyChanged` with private fields and property change notifications.
- Changed `AddAlert` method in `AlertRepository` to use `decimal` for bounds.
- Modified `IStockPageRepository` to return `Stock` instead of `StockPageStock`.
- Simplified stock creation in `StockPageRepository` by removing `quantity` parameter.
- Introduced `UserRepository` for user management with CRUD operations.
- Aligned `AlertService` and `IAlertService` method signatures with updated `Alert` class.
- Enhanced `AlertViewModel` to use `ObservableCollection<Alert>` for dynamic UI updates and added new commands.
- Updated `AlertsView.xaml` for data binding and improved MVVM compliance.
- Refactored `AlertsView.xaml.cs` to remove event handlers in favor of command bindings.
- Added `DialogService` for handling message dialogs and created `IDialogService` interface for better organization and testability.

+++ Repositories/AlertRepository.cs
@@ -68,7 +68,7 @@
                 ?? throw new KeyNotFoundException($"Alert with ID {alertId} not found.");
         }
 
-        public Alert AddAlert(string stockName, string name, int upperBound, int lowerBound, bool toggleOnOff)
+        public Alert AddAlert(string stockName, string name, decimal upperBound, decimal lowerBound, bool toggleOnOff)
         {
             string insertQuery = @"
                 INSERT INTO ALERTS
